Cover Letter Generator/requirements.txt
streamlit
langchain-openai
langchain-core
langchain-community
python-dotenv
pypdf
